Refactor: Introduce breaking changes from codegen
This commit introduces several breaking changes due to code generation and subsequent refactoring attempts.

- **`TrinityRepository`**:
    - Removed explicit `Result` types from `flow` return values.
    - `getAgentStatus` now returns an `AgentResponse` instead of `AgentStatus`.
    - Updated `mapToDomainAgentStatus` to accept `AgentResponse`.
    - Introduced `TODO`s for new helper functions and properties.
    - Updated `parseColor` to use `toColorInt()`.

- **`ConferenceRoomScreen`**:
    - Changed `AgentButton`'s `onClick` parameter to `@Composable () -> Unit`.
    - Added a `TODO` for a new `RowScope.Button` overload.

- **`AIAgentApi`**:
    - The return type of `getAgentStatus` has been changed from `AgentStatus` to `AgentResponse`.

- **Other**:
    - Changed `notificationsEnabledFlow` type to `Any` in `DataStoreManager`.
    - Adjusted imports in `AuraApiService`.
    - Added properties to `AgentResponse`.
