+++ app/src/main/java/dev/aurakai/auraframefx/di/NetworkModule.kt
@@ -11,7 +11,7 @@ import dagger.hilt.android.qualifiers.ApplicationContext
 import dagger.hilt.components.SingletonComponent
 import dev.aurakai.auraframefx.BuildConfig
 import dev.aurakai.auraframefx.network.*
-import dev.aurakai.auraframefx.network.api.ThemeApi
+import dev.aurakai.auraframefx.aura.themes.ThemeApi
 import dev.aurakai.auraframefx.network.api.UserApi
 import dev.aurakai.auraframefx.network.model.*
 import okhttp3.OkHttpClient
