+++ app/src/main/java/dev/aurakai/auraframefx/network/AuraApiService.kt
@@ -4,6 +4,7 @@ package dev.aurakai.auraframefx.network
 
 import AeGenesis.ai.config.AIConfig
 import dev.aurakai.auraframefx.network.api.ThemeApi
+import dev.aurakai.auraframefx.network.api.UserApi
 import retrofit2.http.Body
 import retrofit2.http.GET
 import retrofit2.http.POST
